Naming refactors - rules.py > when_rules.py
