+++ app/questionnaire/router.py
@@ -8,7 +8,7 @@ from app.questionnaire.location import Location
 from app.questionnaire.path_finder import PathFinder
 from app.questionnaire.routing.when_rule_evaluator import WhenRuleEvaluator
 from app.questionnaire.routing_path import RoutingPath
-from app.questionnaire.rules import evaluate_when_rules
+from app.questionnaire.when_rules import evaluate_when_rules
 
 
 class Router:
